Moved helper method to the sole place it was used
